# StudySphere/api/openapi.yaml
openapi: 3.0.3
info:
  title: StudySphere API
  version: 1.0.0
  description: |
    REST API for StudySphere academic planning platform.
    All database interactions must go through this API.
servers:
  - url: https://api.studysphere.com
security:
  - bearerAuth: []
...
paths:
...
  /grades:
    post:
      tags: [Grades]
      summary: Create grade
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GradeCreate'
      responses:
        201:
          description: Created
  /grades/course/{courseId}:
    get:
      tags: [Grades]
      summary: Get grades for a course
      parameters:
        - $ref: '#/components/parameters/courseId'
      responses:
        200:
          description: Grades list
  /grades/{gradeId}:
    patch:
      tags: [Grades]
      summary: Update grade
      parameters:
        - $ref: '#/components/parameters/gradeId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GradeUpdate'
      responses:
        200:
          description: Updated

    delete:
      tags: [Grades]
      summary: Delete grade
      parameters:
        - $ref: '#/components/parameters/gradeId'
      responses:
        204:
          description: Deleted

  # =========================
  # STUDY SESSIONS
  # =========================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    courseId:
      name: courseId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    gradeId:
      name: gradeId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    GradeCreate:
      type: object
      required: [course_id, score, weight]
      properties:
        course_id:
          type: string
        assignment_id:
          type: string
        score:
          type: number
        weight:
          type: number
    GradeUpdate:
      type: object
      properties:
        score:
          type: number
        weight:
          type: number
...